feat(ONB-102): refactor posts
