--- web/src/main/java/co/mobileaction/example/web/service/IPostQueueService.java
@@ -1,11 +1,0 @@
-package co.mobileaction.example.web.service;
-
-/**
- * @author sa
- * @date 17.05.2021
- * @time 17:57
- */
-public interface IPostQueueService
-{
-    void sendPostRequestForAllItems();
-}
