+++ web/src/main/java/co/mobileaction/example/web/service/Post/IPostQueueService.java
@@ -1,0 +1,11 @@
+package co.mobileaction.example.web.service.Post;
+
+/**
+ * @author sa
+ * @date 17.05.2021
+ * @time 17:57
+ */
+public interface IPostQueueService
+{
+    void sendPostRequestForAllItems();
+}
