+++ web/src/main/java/co/mobileaction/example/web/service/Post/IPostResultHandlerService.java
@@ -1,0 +1,13 @@
+package co.mobileaction.example.web.service.Post;
+
+import co.mobileaction.example.common.dto.PostDto;
+
+/**
+ * @author sa
+ * @date 17.05.2021
+ * @time 17:39
+ */
+public interface IPostResultHandlerService
+{
+    void executeMessage(PostDto postDto);
+}
