+++ web/src/test/java/co/mobileaction/example/web/controller/AdminControllerTests.java
@@ -1,6 +1,6 @@
 package co.mobileaction.example.web.controller;
 
-import co.mobileaction.example.web.service.IPostQueueService;
+import co.mobileaction.example.web.service.Post.IPostQueueService;
 import co.mobileaction.example.web.util.SecurityUtils;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
