+++ web/src/test/java/co/mobileaction/example/web/controller/PostControllerTests.java
@@ -1,7 +1,7 @@
 package co.mobileaction.example.web.controller;
 
 import co.mobileaction.example.web.model.Post;
-import co.mobileaction.example.web.service.IPostService;
+import co.mobileaction.example.web.service.Post.IPostService;
 import co.mobileaction.example.web.util.SecurityUtils;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
