+++ web/src/test/java/co/mobileaction/example/web/service/PostServiceTests.java
@@ -2,6 +2,7 @@ package co.mobileaction.example.web.service;
 
 import co.mobileaction.example.web.model.Post;
 import co.mobileaction.example.web.repository.IPostRepository;
+import co.mobileaction.example.web.service.Post.IPostService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
